Added man pages for the BASE package names.
These man pages should pop up whenever someone types ?packageName.db
the aim is for them to have enough information so that the user can
find out 1) what objects this class makes avaialable and 2) something
about the purpose of the class.  The details about where the data came
from etc. are supposed to be in the more detailed man pages for each
provided environment.  To put that information in this package feels
misleading, so this is a change from before.



git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/AnnotationDbi@27913 bc3139a8-67e5-0310-9ffc-ced21a209358