Major change to SQL.R.  Now Bimaps can join across DBs.
This updates AnnotationDbi so that Bimaps can be generated across
SQLite databases from multiple different installed packages.
Providing that the databases are indexed appropriately, there does not
seem to be a huge time penalty to such shenanigans, but there is still
some performance lost when doing this (on the order of a tenth of a
second for most maps tested).  However this will hopefully be offset
by the fact that we will no longer have to have data in all of the
chip packages which is identical to the data in all our org
packages...



git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/AnnotationDbi@35978 bc3139a8-67e5-0310-9ffc-ced21a209358