Added a warning for all arabidopsis chip packages
This warning is to notify users that they will need to use the
toggleProbes() method in order to expose mutliple mappings.  This
change has finally brought the way we use arabidopsis packages in line
with how we use the rest of our annoation packages.  But users still
need to be told about this change, so we are putting a friendly tip in
here when the package 1st loads.  It is wrapped in a
packageStartupMessage() so it can be filtered out if you should find
it annoying.  



git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/AnnotationDbi@41734 bc3139a8-67e5-0310-9ffc-ced21a209358